RAG_Challenge/requirements.txt
# Core Framework
streamlit==1.31.0
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.24

# Vector Store & Database
psycopg2-binary==2.9.9
pgvector==0.2.5
neo4j==5.17.0

# Knowledge Graph
spacy==3.7.4

# Document Processing
pypdf==4.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
unstructured==0.12.4

# NLP & Entity Recognition
transformers==4.38.1
torch==2.2.0

# Utilities
python-dotenv==1.0.1
openai==1.12.0
tiktoken==0.5.2
numpy==1.26.4
pandas==2.2.0

# UI Enhancement
plotly==5.19.0
streamlit-chat==0.1.1